Recurrent map checks, fixes

+++ app/model/connections.js
@@ -1,27 +1,27 @@
 module.exports = function(sequelize, DataTypes) {
   return sequelize.define('connections', {
     id: {
-      type: DataTypes.INTEGER(11),
+      type: DataTypes.INTEGER(),
       allowNull: false,
       primaryKey: true,
       autoIncrement: true
     },
     source: {
-      type: DataTypes.INTEGER(11),
+      type: DataTypes.INTEGER(),
       allowNull: false,
     },
     target: {
-      type: DataTypes.INTEGER(11),
+      type: DataTypes.INTEGER(),
       allowNull: false,
     },
     cost: {
-      type: DataTypes.INTEGER(11),
+      type: DataTypes.INTEGER(),
     },
     reverseCost: {
-      type: DataTypes.INTEGER(11),
+      type: DataTypes.INTEGER(),
     },
     day: {
-      type: DataTypes.INTEGER(11),
+      type: DataTypes.INTEGER(),
       allowNull: false
     }
   }, {
